Add files via upload
Welcome to the prime number wiki!
I made program withe python of Prime number.What this program do? you give it a integer number and it appear to you all the prime number less than or equal the integer number,but with a method of calculation cause of CPU.I hope that you like him.